backend/pyproject.toml
[tool.poetry]
name = "hc-clubsdirectory"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pyairtable = "^1.4.0"
python-dotenv = "^1.0.0"
slack-sdk = "^3.21.3"
geopy = "^2.3.0"
pycountry-convert = "^0.7.2"

[tool.poetry.dev-dependencies]
rich = "^13.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
